requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
dnspython==1.15.0
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2-binary==2.7.5
pymongo==3.7.2
python-dateutil==2.7.5
python-decouple==3.1
python-editor==1.0.3
pytz==2019.2
requests==2.20.0
six==1.11.0
sqlparse==0.3.0
urllib3==1.24.1